Fixed number of decimal places

+++ src/main/java/org/example/Entity/IsoscelesTrapezoid.java
@@ -2,11 +2,11 @@ package org.example.Entity;
 
 public class IsoscelesTrapezoid extends Figure {
 
-    private final double base1;
-    private final double base2;
-    private final double height;
+    private final int base1;
+    private final int base2;
+    private final int height;
 
-    public IsoscelesTrapezoid(String color, double base1, double base2, double height) {
+    public IsoscelesTrapezoid(String color, int base1, int base2, int height) {
         super(color);
         this.base1 = base1;
         this.base2 = base2;
@@ -15,7 +15,7 @@ public class IsoscelesTrapezoid extends Figure {
 
     @Override
     public double getArea() {
-        return (base1+base2) / 2 * height;
+        return (double) (base1 + base2) / 2 * height;
     }
 
     @Override
